+++ src/__init__.py
@@ -1,7 +1,8 @@
 
+#from importlib.metadata import distribution 
+__version__ = '0.2.0.rc6'#distribution('isotools').version
 from .gene import Gene
 from .transcriptome import Transcriptome
 from .splice_graph import SegmentGraph, SegGraphNode
-from ._version import __version__
 
 from ._transcriptome_filter import DEFAULT_GENE_FILTER, DEFAULT_TRANSCRIPT_FILTER, DEFAULT_REF_TRANSCRIPT_FILTER, ANNOTATION_VOCABULARY